Updating to the rtmidi fork dependency changes.
The RtMidi fork we use (https://github.com/remy-j-a-moueza/rtmidi with the
c-api-changes) have been updated.

The impacting change concerns the use of `rtmidi_in_get_message()` that now
expects an allocated message array, to ease the use of the API in C# and
other managed memory languages with no access to the c stdlib `free()`
function but capable of providing a pointer to an array.

We also took the opportunity to:
- update the README.md to tell about the RtMidi fork dependency
- replace `dub.json` tabs to spaces (personal preferences).